配方炉管信息新增！

+++ Demo/Forms/Tube/frmHeating.cs
@@ -44,8 +44,8 @@ namespace Demo.Forms.Tube
         /// <param name="e"></param>
         private void btnOK_Click(object sender, EventArgs e)
         {
-            MyModus.grHMI_TempSV =  Convert.ToDouble(txtTempSV.Text);
-            MyModus.grHMI_TempRump = Convert.ToDouble(txtTempRump.Text);
+            MyModus.grHMI_TempSV = (float)Convert.ToDouble(txtTempSV.Text);
+            MyModus.grHMI_TempRump = (float)Convert.ToDouble(txtTempRump.Text);
             MyModus.gbHMI_Temp_Enable = true;
             this.Close();
         }
